WIP - Converting tests to approval tests
