+++ exercise/kotlin/day14/src/test/kotlin/children/ChildMapperApprovalTests.should map X5T78 to Child for a boy.approved.txt
@@ -1,0 +1,1 @@
+Child(id=62e0c30c-74da-40f4-b951-0c56837d3704, firstName=Bob, middleName=null, lastName=Brown, birthCity=Paradise, birthDate=01/09/2021, gender=Boy, address=Address(number=9, street=Oak Street, city=Paradise, countryId=98988))
